requirements_cpu.txt
numpy>=1.16
pathlib
scipy==1.3  # 1.2 is needed for baselines stuff
Pillow
gym[atari]
pyyaml
ipdb
termcolor
tensorflow==1.12
opencv-python
gin-config==0.1.4
joblib
cloudpickle
pytest
matplotlib
tensorpack
attrs
tee
pyzmq
mpi4py
seaborn
numba
deprecated
munch
pathlib2
moviepy
pygame
